Bug #987996 dcam&isp interrupt debug interface
[root cause  ] sharkl5 camera debug
[changes     ] kernel_module/core/isp2.6
[side effects] no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:ums510-1h10
[test case]:camera preview/capture
[reviewers   ] victor.xu
[change_type ] bringup --> sprd_chip_support
[tag_product ] specific

+++ kernel_module/core/isp2.6/inc/dcam_core.h
@@ -179,7 +179,7 @@ struct dcam_pipe_dev {
 	uint32_t irq;
 	atomic_t state;// TODO: use mutex to protect
 
-	uint64_t frame_index;
+	uint32_t frame_index;
 	uint32_t enable_slowmotion;
 	uint32_t slowmotion_count;
 	uint32_t helper_enabled;
